# Performance gains by moving IndexGet and IndexSet into Object interface 