--- a/objects/array_iterator.go
+++ b/objects/array_iterator.go
@@ -1,9 +1,8 @@
 package objects
 
-import "github.com/d5/tengo/compiler/token"
-
 // ArrayIterator is an iterator for an array.
 type ArrayIterator struct {
+	ObjectImpl
 	v []Object
 	i int
 	l int
@@ -16,17 +15,6 @@ func (i *ArrayIterator) TypeName() string {
 
 func (i *ArrayIterator) String() string {
 	return "<array-iterator>"
-}
-
-// BinaryOp returns another object that is the result of
-// a given binary operator and a right-hand side object.
-func (i *ArrayIterator) BinaryOp(op token.Token, rhs Object) (Object, error) {
-	return nil, ErrInvalidOperator
-}
-
-// IsFalsy returns true if the value of the type is falsy.
-func (i *ArrayIterator) IsFalsy() bool {
-	return true
 }
 
 // Equals returns true if the value of the type
@@ -54,14 +42,4 @@ func (i *ArrayIterator) Key() Object {
 // Value returns the value of the current element.
 func (i *ArrayIterator) Value() Object {
 	return i.v[i.i-1]
-}
-
-// IndexGet returns an element at a given index.
-func (i *ArrayIterator) IndexGet(index Object) (Object, error) {
-	return nil, ErrNotIndexable
-}
-
-// IndexSet sets an element at a given index.
-func (i *ArrayIterator) IndexSet(index, value Object) error {
-	return ErrNotIndexAssignable
 }