--- a/objects/builtin_function.go
+++ b/objects/builtin_function.go
@@ -45,3 +45,13 @@ func (o *BuiltinFunction) Equals(x Object) bool {
 func (o *BuiltinFunction) Call(args ...Object) (Object, error) {
 	return o.Value(args...)
 }
+
+// IndexGet returns an element at a given index.
+func (o *BuiltinFunction) IndexGet(index Object) (Object, error) {
+	return nil, ErrNotIndexable
+}
+
+// IndexSet sets an element at a given index.
+func (o *BuiltinFunction) IndexSet(index, value Object) error {
+	return ErrNotIndexAssignable
+}