--- a/objects/builtin_type_checks.go
+++ b/objects/builtin_type_checks.go
@@ -174,8 +174,7 @@ func builtinIsCallable(args ...Object) (Object, error) {
 		return nil, ErrWrongNumArguments
 	}
 
-	switch args[0].(type) {
-	case *CompiledFunction, *Closure, Callable: // BuiltinFunction is Callable
+	if args[0].CanCall() {
 		return TrueValue, nil
 	}
 
@@ -187,7 +186,7 @@ func builtinIsIterable(args ...Object) (Object, error) {
 		return nil, ErrWrongNumArguments
 	}
 
-	if iter := args[0].Iterate(); iter != nil {
+	if args[0].CanIterate() {
 		return TrueValue, nil
 	}