--- a/objects/char.go
+++ b/objects/char.go
@@ -6,6 +6,7 @@ import (
 
 // Char represents a character value.
 type Char struct {
+	ObjectImpl
 	Value rune
 }
 
@@ -116,14 +117,4 @@ func (o *Char) Equals(x Object) bool {
 	}
 
 	return o.Value == t.Value
-}
-
-// IndexGet returns an element at a given index.
-func (o *Char) IndexGet(index Object) (Object, error) {
-	return nil, ErrNotIndexable
-}
-
-// IndexSet sets an element at a given index.
-func (o *Char) IndexSet(index, value Object) error {
-	return ErrNotIndexAssignable
 }