--- a/objects/closure.go
+++ b/objects/closure.go
@@ -6,6 +6,7 @@ import (
 
 // Closure represents a function closure.
 type Closure struct {
+	ObjectImpl
 	Fn   *CompiledFunction
 	Free []*ObjectPtr
 }
@@ -42,14 +43,4 @@ func (o *Closure) IsFalsy() bool {
 // is equal to the value of another object.
 func (o *Closure) Equals(x Object) bool {
 	return false
-}
-
-// IndexGet returns an element at a given index.
-func (o *Closure) IndexGet(index Object) (Object, error) {
-	return nil, ErrNotIndexable
-}
-
-// IndexSet sets an element at a given index.
-func (o *Closure) IndexSet(index, value Object) error {
-	return ErrNotIndexAssignable
 }