--- a/objects/float.go
+++ b/objects/float.go
@@ -144,3 +144,13 @@ func (o *Float) Equals(x Object) bool {
 
 	return o.Value == t.Value
 }
+
+// IndexGet returns an element at a given index.
+func (o *Float) IndexGet(index Object) (Object, error) {
+	return nil, ErrNotIndexable
+}
+
+// IndexSet sets an element at a given index.
+func (o *Float) IndexSet(index, value Object) error {
+	return ErrNotIndexAssignable
+}