--- a/objects/float.go
+++ b/objects/float.go
@@ -9,6 +9,7 @@ import (
 
 // Float represents a floating point number value.
 type Float struct {
+	ObjectImpl
 	Value float64
 }
 
@@ -143,14 +144,4 @@ func (o *Float) Equals(x Object) bool {
 	}
 
 	return o.Value == t.Value
-}
-
-// IndexGet returns an element at a given index.
-func (o *Float) IndexGet(index Object) (Object, error) {
-	return nil, ErrNotIndexable
-}
-
-// IndexSet sets an element at a given index.
-func (o *Float) IndexSet(index, value Object) error {
-	return ErrNotIndexAssignable
 }