--- a/objects/iterable.go
+++ b/objects/iterable.go
@@ -1,7 +1,0 @@
-package objects
-
-// Iterable represents an object that has iterator.
-type Iterable interface {
-	// Iterate should return an Iterator for the type.
-	Iterate() Iterator
-}