--- a/objects/string.go
+++ b/objects/string.go
@@ -90,6 +90,11 @@ func (o *String) IndexGet(index Object) (res Object, err error) {
 	return
 }
 
+// IndexSet sets an element at a given index.
+func (o *String) IndexSet(index, value Object) error {
+	return ErrNotIndexAssignable
+}
+
 // Iterate creates a string iterator.
 func (o *String) Iterate() Iterator {
 	if o.runeStr == nil {