… docs/objects.md
… objects/object_impl.go
… objects/string_iterator.go
… objects/time.go
… objects/undefined.go
… objects/user_function.go
… runtime/vm.go
… runtime/vm_indexable_test.go
… runtime/vm_iterable_test.go
… script/script_bench_test.go
… script/script_custom_test.go