Removing type assertion to `Indexable` and `IndexAssignable` interfaces improves overall `IndexGet` and `IndexSet` performance significantly at the downside of having to implement these methods for all objects. I think it's a good tradeoff.

BEFORE
```
BenchmarkArrayIndex-28                      5000        298004 ns/op
BenchmarkArrayIndexComparison-28           10000        107746 ns/op
```

AFTER
```
BenchmarkArrayIndex-28                     10000        216165 ns/op
BenchmarkArrayIndexComparison-28           10000        108115 ns/op
```